pyproject.toml
[project]
name = "news-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.3"
dependencies = [
    "fastapi[standard]>=0.128.7",
    "beanie>=2.0.0",
    "motor>=3.0.0",
    "pymongo>=4.10",
    "openai>=1.55",
    "python-dotenv>=1.0",
    "pydantic-settings>=2.0.0",
    "haversine>=2.9.0",
    "httpx>=0.27",
    "redis>=5.0",
    "typer>=0.15",
    "pyyaml>=6.0",
]

[dependency-groups]
dev = ["ruff>=0.8"]

[tool.ruff]
target-version = "py313"
line-length = 88

[tool.ruff.format]
quote-style = "double"
